Improve top header
